yellow_requirements.txt
python-dotenv==0.19.0
redis==3.5.3
retrying==1.3.3
selenium==3.141.0
six==1.16.0
urllib3==1.26.7
